# Conflicts:
#	Maze.txt
#	frontend/src/components/CreateLobbyForm.vue
#	frontend/src/stores/Lobby/lobbiesstore.ts
#	frontend/src/stores/Player/IPlayerDTD.ts
#	frontend/src/stores/Player/IPlayerInitDTD.ts
#	frontend/src/stores/Snack/ISnackDTD.ts
#	frontend/src/stores/gameMapStore.ts
#	frontend/src/views/GameView.vue
#	frontend/src/views/LobbyViews/LobbyListView.vue
#	frontend/src/views/LobbyViews/LobbyView.vue
#	scripts/Maze.py
#	src/main/java/de/hsrm/mi/swt/snackman/SnackmanApplication.java
#	src/main/java/de/hsrm/mi/swt/snackman/configuration/GameConfig.java
#	src/main/java/de/hsrm/mi/swt/snackman/controller/GameMap/GameMapController.java
#	src/main/java/de/hsrm/mi/swt/snackman/controller/GameMap/SnackManInitDTO.java
#	src/main/java/de/hsrm/mi/swt/snackman/controller/Lobby/LobbyController.java
#	src/main/java/de/hsrm/mi/swt/snackman/controller/Lobby/PlayerClientController.java
#	src/main/java/de/hsrm/mi/swt/snackman/controller/PlayerMovement/PlayerMovementController.java
#	src/main/java/de/hsrm/mi/swt/snackman/controller/PlayerMovement/SnackManFrontendDTO.java
#	src/main/java/de/hsrm/mi/swt/snackman/controller/PlayerMovement/SnackManPositionDTO.java
#	src/main/java/de/hsrm/mi/swt/snackman/entities/lobby/Lobby.java
#	src/main/java/de/hsrm/mi/swt/snackman/entities/map/GameMap.java
#	src/main/java/de/hsrm/mi/swt/snackman/entities/mapObject/snack/SnackType.java
#	src/main/java/de/hsrm/mi/swt/snackman/entities/mobileObjects/Mob.java
#	src/main/java/de/hsrm/mi/swt/snackman/entities/mobileObjects/eatingMobs/Chicken/Chicken.java
#	src/main/java/de/hsrm/mi/swt/snackman/entities/mobileObjects/eatingMobs/Chicken/Direction.java
#	src/main/java/de/hsrm/mi/swt/snackman/entities/mobileObjects/eatingMobs/EatingMob.java
#	src/main/java/de/hsrm/mi/swt/snackman/entities/mobileObjects/eatingMobs/SnackMan.java
#	src/main/java/de/hsrm/mi/swt/snackman/messaging/EventType.java
#	src/main/java/de/hsrm/mi/swt/snackman/messaging/FrontendMessageService.java
#	src/main/java/de/hsrm/mi/swt/snackman/services/LobbyManagerService.java
#	src/main/java/de/hsrm/mi/swt/snackman/services/MapService.java
#	src/test/java/de/hsrm/mi/swt/snackman/entities/mobileObjects/eatingMobs/SnackManTest.java
#	src/test/java/de/hsrm/mi/swt/snackman/services/MapServiceTest.java

+++ src/main/java/de/hsrm/mi/swt/snackman/entities/mobileObjects/Mob.java
@@ -32,7 +32,7 @@ public abstract class Mob {
      * @param speed      speed of the mob
      * @param radius     size of the mob
      */
-    public Mob(GameMap gameMap, int speed, double radius) {
+    public Mob(GameMap gameMap, double speed, double radius) {
         this.gameMap = gameMap;
         this.speed = speed;
         this.radius = radius;
@@ -58,7 +58,7 @@ public abstract class Mob {
      * @param posY       y-spawn-positon
      * @param posZ       z-spawn-position
      */
-    public Mob(GameMap gameMap, int speed, double radius, double posX, double posY, double posZ) {
+    public Mob(GameMap gameMap, double speed, double radius, double posX, double posY, double posZ) {
         this(gameMap, speed, radius);
 
         spawn = new Vector3d(posX, posY, posZ);
